Added frontend ui

+++ backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 import asyncHandler from 'express-async-handler'
 import Goal from '../models/goalModel.js'
+import User from '../models/userModel.js'
 
 // @desc Get goals
 // @route GET /api/goals
 // @access private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
 
     res.status(200).json(goals)
 })
@@ -21,6 +22,7 @@ const createGoal = asyncHandler(async (req, res) => {
 
     const goal = await Goal.create({
         text: req.body.text,
+        user: req.user.id,
     })
 
     res.status(200).json(goal)
@@ -36,6 +38,20 @@ const updateGoal = asyncHandler(async (req, res) => {
         throw new Error('Goal not found')
     }
 
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure the logged in user matches goal user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     const updGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
     })
@@ -47,15 +63,27 @@ const updateGoal = asyncHandler(async (req, res) => {
 // @route DELETE /api/goals/:id
 // @access private
 const deleteGoal = asyncHandler(async (req, res) => {
-    const goal = Goal.findById(req.params.id)
-    if(!goal) {
+    const goal = await Goal.findById(req.params.id)
+    if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
     }
 
+    // Check for user
+    if (!req.user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure the logged in user matches goal user
+    if (goal.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     await goal.deleteOne()
 
-    res.status(200).json({id: req.params.id})
+    res.status(200).json({ id: req.params.id })
 })
 
 export {
